Add error checks to monitor and simulation

+++ source/philo/include/modules/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: yademirk <yademirk@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/04 16:16:51 by yademirk          #+#    #+#             */
-/*   Updated: 2026/03/14 20:24:49 by yademirk         ###   ########.fr       */
+/*   Updated: 2026/03/15 22:31:38 by yademirk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
  */
 int		should_philo_continue(t_philosopher *philo);
 
-void	join_philosophers(t_philosopher *philos, int count);
+void	join_philosophers(t_philosopher *philos, size_t count);
 
 /**
  * @brief Initializes the philosophers of a table,
@@ -29,6 +29,6 @@ void	join_philosophers(t_philosopher *philos, int count);
  * @return 0 on error, otherwise returns the number of
  * successfully started threads.
  */
-int		start_philosophers(t_table *table, int count);
+size_t	start_philosophers(t_table *table, size_t count);
 
 #endif
